moving files into user interface framework
