+++ modules/SwiftAlgorithmsUserInterface/SwiftAlgorithmsUserInterface/Classes/Reusable.swift
@@ -1,8 +1,8 @@
-protocol Reusable {
+public protocol Reusable {
     static var reuseIdentifier: String { get }
 }
 
-extension Reusable where Self: UIView {
+public extension Reusable where Self: UIView {
     static var reuseIdentifier: String {
         return String(describing: self)
     }
